change layout to match CDR, fix interrupts with skytraq
